refactor: merge profile info into Your Data section, remove ProfileInfo
Avatar, name, email, and connected providers now display under the
"Your Data" heading alongside download and delete actions.
